requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
flask==3.0.3
werkzeug==3.0.3

# ── Gemini ─────────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ── Streamlit frontend ─────────────────────────────────────────────────────────
streamlit==1.36.0
requests==2.32.3
